requirements-dev.txt
fake-bpy-module-2.82
setuptools
setuptools-rust
blender.distutils
black
